refactor: modularize stockbot controller
